+++ CHANGES.md
@@ -9,6 +9,7 @@ To be released.
  -  Added `Bencodex.Types.Binary.FromHex()` static method.  [[#69]]
  -  Added `Bencodex.Types.Binary.FromBase64()` static method.  [[#69]]
  -  Added `Bencodex.Types.Binary.ToHex()` method.  [[#69]]
+ -  Added `Bencodex.Types.Binary.ToBase64()` method.  [[#69]]
 
 [#69]: https://github.com/planetarium/bencodex.net/pull/69